Checkpoint - moved update compoonent to components and renamed to update-container
