--- src/app/modules/msp-update/components/groups/groups.component.html
@@ -1,80 +1,0 @@
-<common-page-framework layout='blank'>
-  <h3>MSP Group Numbers</h3>
-  <p>MSP Group Numbers administered through MSP Direct on behalf of the Organization.</p>
-  <hr>
-
-  <common-page-section layout='noTips'>
-    <div class="row">
-      <button type='button' class='btn btn-secondary col mr-4' (click)="addMspGroup()">Add MSP Group</button>
-      <button type='button' class='btn btn-secondary col mx-2' (click)="removeMspGroup()">Remove MSP Group</button>
-      <button type='button' class='btn btn-secondary col ml-4' (click)="updateMspGroupAdmin()">Update administration of MSP Group</button>
-    </div>
-  </common-page-section>
-
-  <common-page-section layout='noTips' *ngIf="showAddMspGrp" class="pt-4">
-      <form [formGroup]="addFg">
-        <div formArrayName="addGrpItems">
-          <div *ngFor="let item of addGrpArray.controls; let idx = index"
-              [formGroupName]="idx"
-              class="pt-2">
-            <common-xicon-button label="Add Msp Group" (clickEvent)="deleteAddGrpItem(idx)"></common-xicon-button>
-            <h4 class='border-bottom'>Add new MSP Group #{{idx + 1}}</h4>
-
-            <div class="form-group col-sm-6 p-sm-0">
-              <sitereg-msp-group-no labelforId="goupNo_{{idx}}"
-                                    formControlName="groupNo"></sitereg-msp-group-no>
-              <common-radio [radioLabels]="radioBtnLabels"
-                            formControlName="thirdPartyAdmin"
-                            label="Will this group be administered by a Third Party Administrator?"></common-radio>
-            </div>
-          </div>
-        </div>
-    </form>
-  </common-page-section>
-
-  <common-page-section layout='noTips' *ngIf="showRemoveMspGrp" class="pt-4">
-    <form [formGroup]="removeFg">
-        <div formArrayName="removeGrpItems">
-          <div *ngFor="let item of removeGrpArray.controls; let idx = index"
-               [formGroupName]="idx"
-               class="pt-2">
-            <common-xicon-button label="Remove Msp Group" (clickEvent)="deleteRemoveGrpItem(idx)"></common-xicon-button>
-            <h4 class='border-bottom'>Remove new MSP Group #{{idx + 1}}</h4>
-
-            <div class="form-group col-sm-6 p-sm-0">
-              <sitereg-msp-group-no labelforId="goupNo_{{idx}}"
-                                    formControlName="groupNo"></sitereg-msp-group-no>
-            </div>
-          </div>
-        </div>
-    </form>
-  </common-page-section>
-
-  <common-page-section layout='noTips' *ngIf="showUpdateMspGrpAdmin" class="pt-4">
-    <form [formGroup]="updateFg">
-        <div formArrayName="updateGrpItems">
-          <div *ngFor="let item of updateGrpArray.controls; let idx = index"
-              [formGroupName]="idx"
-              class="pt-2">
-            <common-xicon-button label="Update administration of MSP Group" (clickEvent)="deleteUpdateGrpItem(idx)"></common-xicon-button>
-            <h4 class='border-bottom'>Update administration of MSP Group #{{idx + 1}}</h4>
-
-            <div class="form-group col-sm-6 p-sm-0">
-              <sitereg-msp-group-no labelforId="goupNo_{{idx}}"
-                                    formControlName="groupNo"></sitereg-msp-group-no>
-              <common-radio [radioLabels]="radioBtnLabels"
-                            formControlName="thirdPartyAdmin"
-                            label="Is the group managed bya Third Party Administrator?"></common-radio>
-            </div>
-          </div>
-        </div>
-    </form>
-  </common-page-section>
-
-</common-page-framework>
-
-<common-form-action-bar (btnClick)="continue()"
-                        widthOption='extra-width-mobile-only'
-                        [submitLabel]="buttonLabel"
-                        [canContinue]='canContinue()'>
-</common-form-action-bar>
