+++ src/app/modules/msp-update/routing/routes.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { GlobalConfigService } from '@shared/services/global-config.service';
-import { MspDirectUpdateComponent } from '../pages/update/update.component';
+import { MspDirectUpdateComponent } from '../components/update-container/update-container.component'
 import { routes } from './routes';
