+++ src/app/shared/components/h2-with-print/h2-with-print.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 
 @Component({
-  selector: 'sitereg-msp-h3-with-print',
+  selector: 'sitereg-msp-h2-with-print',
   templateUrl: './h2-with-print.component.html',
   styleUrls: ['./h2-with-print.component.scss']
 })
