+++ src/version.GENERATED.ts
@@ -1,9 +1,9 @@
 // DO NOT DELETE OR APP WILL FAIL TO COMPILE! Generated from version.js
-export const gitCommit = '3aaa803';
-export const buildTime = '6/24/2019 at 10:44:17 AM';
+export const gitCommit = '16dd7e4';
+export const buildTime = '6/24/2019 at 12:01:59 PM';
 /** App version retrieved from package.json. */
 export const projectVersion = '0.0.0';
 /** Human readable message  */
-export const message = 'Canonical Version: 3aaa803 - 6/24/2019 at 10:44:17 AM (v0.0.0) \nhttps://github.com/bcgov/moh-sitereg/commits/3aaa803';
+export const message = 'Canonical Version: 16dd7e4 - 6/24/2019 at 12:01:59 PM (v0.0.0) \nhttps://github.com/bcgov/moh-sitereg/commits/16dd7e4';
 /** If true, other values should be present. */
 export const success = true;
